# common-platform-product-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: common-platform-product-mgmt
  description: Product Management Core Application
  contact:
    name: Catalis Development
    email: dev@catalis.com
  version: 1.0.0
servers:
  - url: 'http://core.catalis.vc/products'
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  '/api/v1/products/{productId}/features/{featureId}':
    get:
      tags:
        - Product Feature
      summary: Get Product Feature by ID
      description: >-
        Retrieve a specific product feature using its unique identifier,
        ensuring it matches the product.
      operationId: getFeature
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          schema:
            type: integer
            format: int64
        - name: featureId
          in: path
          description: Unique identifier of the feature
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved the product feature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductFeatureDTO'
        '404':
          description: Product feature not found
    put:
      tags:
        - Product Feature
      summary: Update Product Feature
      description: Update an existing feature associated with the specified product.
      operationId: updateFeature
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          schema:
            type: integer
            format: int64
        - name: featureId
          in: path
          description: Unique identifier of the product feature to update
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductFeatureDTO'
        required: true
      responses:
        '200':
          description: Product feature updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductFeatureDTO'
        '404':
          description: Product feature not found
    delete:
      tags:
        - Product Feature
      summary: Delete Product Feature
      description: Remove an existing feature from the specified product.
      operationId: deleteFeature
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          schema:
            type: integer
            format: int64
        - name: featureId
          in: path
          description: Unique identifier of the product feature to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Product feature deleted successfully
        '404':
          description: Product feature not found
...
  '/api/v1/products/{productId}':
    get:
      tags:
        - Product
      summary: Get Product by ID
      description: Retrieve a specific product by its unique identifier.
      operationId: getProduct
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved the product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '404':
          description: Product not found
    put:
      tags:
        - Product
      summary: Update Product
      description: Update the information of an existing product by its unique identifier.
      operationId: updateProduct
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product to update
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDTO'
        required: true
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '404':
          description: Product not found
    delete:
      tags:
        - Product
      summary: Delete Product
      description: Remove an existing product by its unique identifier.
      operationId: deleteProduct
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Product deleted successfully
        '404':
          description: Product not found
...
  '/api/v1/products/{productId}/features/filter':
    post:
      tags:
        - Product Feature
      summary: List Product Features
      description: >-
        Retrieve a paginated list of all features associated with the specified
        product.
      operationId: getAllFeatures
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaginationRequest'
        required: true
      responses:
        '200':
          description: Successfully retrieved the list of product features
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseProductFeatureDTO'
        '404':
          description: No features found for the specified product
  '/api/v1/products/{productId}/features':
    post:
      tags:
        - Product Feature
      summary: Create Product Feature
      description: Create a new feature linked to the specified product.
      operationId: createFeature
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductFeatureDTO'
        required: true
      responses:
        '200':
          description: Product feature created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductFeatureDTO'
        '400':
          description: Invalid feature data provided
...
  /api/v1/products/filter:
    post:
      tags:
        - Product
      summary: List Products
      description: Retrieve a paginated list of products.
      operationId: getAllProducts
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaginationRequest'
        required: true
      responses:
        '200':
          description: Successfully retrieved the list of products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseProductDTO'
        '404':
          description: No products found
...
components:
  schemas:
...
    ProductFeatureDTO:
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        productFeatureId:
          type: integer
          format: int64
          readOnly: true
        productId:
          type: integer
          format: int64
        featureName:
          type: string
        featureDescription:
          type: string
        featureType:
          type: string
          enum:
            - STANDARD
            - OPTIONAL
            - PREMIUM
        isMandatory:
          type: boolean
...
    ProductDTO:
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        productId:
          type: integer
          format: int64
          readOnly: true
        productSubtypeId:
          type: integer
          format: int64
        productType:
          type: string
          enum:
            - FINANCIAL
            - NON_FINANCIAL
        productName:
          type: string
        productCode:
          type: string
        productDescription:
          type: string
        productStatus:
          type: string
          enum:
            - PROPOSED
            - ACTIVE
            - RETIRED
        launchDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: 'The direction of sorting, either ASC or DESC.'
          example: DESC
          default: DESC
      description: >-
        Represents a pagination request for retrieving paginated results,
        including page number, size, sort field, and direction.
...
    PaginationResponseProductFeatureDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/ProductFeatureDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: 'The current page number, typically zero-based.'
          format: int32
      description: >-
        Represents a paginated response containing a list of items and
        pagination metadata.
...
    PaginationResponseProductDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/ProductDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: 'The current page number, typically zero-based.'
          format: int32
      description: >-
        Represents a paginated response containing a list of items and
        pagination metadata.
...